pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prague-europe-weather-pipeline"
version = "0.1.0"
description = "Batch pipeline for ingesting and modeling European weather data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "jinja2>=3.1.4",
    "requests>=2.32.0",
    "psycopg[binary]>=3.1.18",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0"
]

[project.scripts]
prague-weather = "pipeline.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "black>=24.0.0",
  "httpx>=0.27.0"
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pipeline = [
    "dashboard/templates/*.html",
    "dashboard/static/*.css",
    "dashboard/static/*.js",
]
